+++ packages/components/src/dropdown-menu-v2/context.tsx
@@ -1,0 +1,14 @@
+/**
+ * WordPress dependencies
+ */
+import { createContext } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import type { DropdownMenuContext as DropdownMenuContextType } from './types';
+
+export const DropdownMenuContext = createContext<
+	DropdownMenuContextType | undefined
+>( undefined );
+DropdownMenuContext.displayName = 'DropdownMenuV2.Context';
